scheme: move IsAllowedKeyType() out of schemeshard
Move `IsAllowedKeyType()` from `yd/core/tx/schemeshard/schemeshard_utils.h` to
`ydb/core/scheme/scheme_tablerefs.h`.
Stop using private header `schemeshard_utils.h` as public.
