+++ ydb/core/kqp/opt/physical/ya.make
@@ -15,6 +15,7 @@ SRCS(
 )
 
 PEERDIR(
+    ydb/core/scheme
     ydb/core/kqp/common
     ydb/core/kqp/opt/physical/effects
     ydb/library/yql/dq/common
